requirements.txt
flask
flask-limiter
flask-cors
requests
python-dotenv
gunicorn
pypdf
